Split character builder build look section
